Make fletch build for Android using ndk-build.
Before this becomes really useful we need to define a public API to
code against. I'll get started on that next and make sure that we can
actually use this library. For now it is useful to just have the build
script in here to verify that we haven't broken the build for
Android.

R=kasperl@google.com, ajohnsen@google.com
BUG=

Review URL: https://codereview.chromium.org/859973005