Add API to add default shared libraries for the foreign function API.
This is useful on Android where we will probably end up with a
shared library that contains all Dart service jni code. The Android
developer who loads that library knows what it is called. In order
to not have to have that knowledge in the Dart service implementation
it is exposed through the API so it can be added when we actually
have the knowledge.

R=kasperl@google.com
BUG=

Review URL: https://codereview.chromium.org/860063003