First steps towards multiple arguments for service methods.
The idea is to let the service-specific generated code handle the
allocation and deallocation of the request structs -- and let the
generic code in service_api_impl.cc store the necessary bits in
the first 32 bytes of such structs. The rest of the struct is
for arguments and results and should only be handled from generated
code.

R=ager@google.com
BUG=

Review URL: https://codereview.chromium.org/882943003