Support string type in service idl descriptions.
This is not pretty; so take this as a request for comments as well.

I decided to make string a primitive type during compilation, not sure
that was a good move.

At this point I am completely ignoring string encodings.

You get a copy of the string on every invocation of the getter for
a string in a struct.

I'll clean this up a bit tomorrow. Let me know what you think. Ian,
could something like this be turned into something useful for you?

R=zerny@google.com, kasperl@google.com

Review URL: https://codereview.chromium.org/995423002