Record fixed size of instances in number of pointers instead of number of bytes.
When reading Functions from the snapshot always round up the
bytecode size to kPointerSize.

This allow reading in a ia32 snapshot on x64 if we overallocate
the initial heap by a factor of two. I haven't done that in this
change. Instead I will write two sizes in the snapshot in an
upcoming change and chose the right one based on the current
platform.

R=kasperl@google.com

Review URL: https://codereview.chromium.org/1012753002