Implement switch.
This also adds two new bytecodes: PopAndBranch{Back}Long. They are used
for break/continue, where the pop is a side-effect of taking the branch.
This makes it easier to generate bytecode for the stack-walker, as there
are no side-effect pop's related to break/continue anymore.

BUG=
R=zerny@google.com

Review URL: https://codereview.chromium.org/1033273004