Fix broken GNUMakefile: Building fletch requires GYP_CROSSCOMPILE to be set
If one removes the 'out' directory and runs 'make' it will automatically run gyp
as well. But as opposed to the gclient runhook, it was missing the
GYP_CROSSCOMPILE environment variable which caused the generated ninja files to
be broken (since gyp will not emit ninja files for host and target toolset
without the GYP_CROSSCOMPILE being set, which causes circular dependencies in
the ninja build targets, which makes one unable to build).

R=ahe@google.com

Review URL: https://codereview.chromium.org/1091553004