Fix printing of locals in the debugger.
The bytecode pointer for a frame is the return address and is
therefore not the current location in the code. Therefore, when
looking up scope information we have to search for bytecode
pointer minus one. We already do this correctly for other
lookups.

It would have been nice to record the correct bytecode
pointer for a frame. However, that would require us to know the
size of the previous bytecode which we can't unless we iterate
from the top.

R=zerny@google.com

Review URL: https://codereview.chromium.org/1070283003