Keep a set of currently waiting sockets, in a static variable, to keep a reference to them.
Sockets has a pointer to a channel that then again has a pointer to the
receiving thread. While the socket is waiting, it's possible to only
have a reference to it indirectly through the eventhandler's port. Since
the GC cannot look into tracing elements in the eventhandler, a set is
used.

BUG=
R=ager@google.com

Review URL: https://codereview.chromium.org/1105463003