Add support for events/methods on nodes.
This CL adds a simple setup for resource management. For now it supports
event handlers that are specified via a method signature on node
declarations in an immi file.

For example:

  node Foo { void bar(T1 p1, ..., Tn pn); }

Declares a dispatchable event on Foo nodes.

In Dart this is populated as other fields but with a function-valued handler:

  new Foo(bar: (T1 p1, ..., Tn pn) { });

From objective-c this is invoked as:

  [someFoo dispatchBarP1:? ... pn:?];

This fires an async dispatch back to the Dart handler.

The resource manager keeps track of which handlers are active within the
presentation graph at any particular point (with 'serialize' denoting
the commit to a new current graph).

The immi compiler needs substantial cleanup, but I'll postpone that
until a bit more of the structure is in place.

R=ager@google.com
BUG=

Review URL: https://codereview.chromium.org/1106483004