Add debugger test support.
Add a fletchd runtime configuration that is enabled by default. Only
the debugger tests are run in that configuration. The fletchd runtime
configuration uses an output diffing version of the VmCommand and
VmCommandOutput.

I think this is the easiest way to support this at this point, but
other suggestions are welcome.

R=kustermann@google.com, kasperl@google.com

Review URL: https://codereview.chromium.org/1133053002