Debugger: start hiding internal frames.
Parameter stubs are not interesting to the end user. They just reorder
arguments. Hide them in the debugger by default. Allow the user to
toggle the hiding on and off. This is particularly useful if stepping
at the bytecode level. We could consider toggling internal frames
on by default if the user explicitly steps at bytecode level.

Add a test for accessors as well. We do not step into compiler
generated accessors that just reads/writes a field. I think that is
the right behavior.

R=zerny@google.com, kasperl@google.com

Review URL: https://codereview.chromium.org/1130253009