Unfold the program when pushing a new function.
If the program is not unfolded we end up with a heap that contains both
folded and unfolded function which is highly unexpected. :)

(I did not notice this when fixing Unfolding for committing patch sets
because PopToMap didn't work. Therefore the new function that we pushed
was immediately garbage and the Unfolding operation didn't encounter it.
I then fixed PopToMap but forgot to rerun the incremental compilation
test.)

R=kasperl@google.com, ahe@google.com

Review URL: https://codereview.chromium.org/1141713002