Add a navigation drawer to the github sample app.
This CL just adds a navigation drawer with static content and no Dart
counterpart. The drawer is created using an external MMDrawerController
library that is managed via the CocoaPods package system for OSX/iOS.

This CL adds a podfile with a dependency for MMDrawerController and
updates the xcode configuration to use a workspace.

After this change users need to install and run the cocoapods packaging
tool prior to building the github sample.

R=zarah@google.com, kasperl@google.com
BUG=

Review URL: https://codereview.chromium.org/1133423002