Introduce explicit PrepareForChanges command to the wire protocol.
All changes to a live system needs to send commands of the form:

PrepareForChanges()
...  // build changes here.
CommitChanges(numberOfChanges)

When the VM receives the PrepareForChanges command it can ensure
that the program is in a valid state to receive changes. Currently
the only requirement is that the program is unfolded.

R=kasperl@google.com, ahe@google.com

Review URL: https://codereview.chromium.org/1132223006