Add the option of running debugger tests for a list of commands.
The list of commands is provided is a '// FletchDebuggerCommands='
comment in the test file.

This is far from perfect, but it leverages test.dart infrastructure
that is already in place to be able to write tests of the full
debugger functionality. As an example, this changelist contains the
first test that verifies that 'step over' works.

R=ahe@google.com, kasperl@google.com

Review URL: https://codereview.chromium.org/1133473010