Abstract the interaction and use of the immi service layer.
This CL introduces ObjC abstractions to support the use of multiple
presentation graphs. The ObjC and Dart implementation to actually
support multiple graphs will follow in subsequent CLs.

The ImmiService class will provide access to the shared service
resources, such as registered presenter objects. With the current
service layer this could be defined as a singleton, but we don't to
avoid introducing implicit dependencies on the shared resources.

The ImmiRoot class abstracts the notion of a presentation graph. It
will allow us to represent multiple presentation graphs in the same
application and will provide access to resources local to a particular
graph.

The RootPresenter protocol defines the interface for objects that can
interpret a full presentation graph. To avoid manually defining
implementations of all navigation controllers in an app, we define a
'Category' that extends the UINavigationController class to satisfy the
RootPresenter as a proxy of the top-most view controller.

R=ager@google.com, zarah@google.com
BUG=

Review URL: https://codereview.chromium.org/1143233002