Implement support for multiple independent presentation graphs.
To associate a presenter (defined in Dart) with an interpreting object
(defined in the native layer), presenters can be registered by-name from
both sides. Doing so associates the two objects with a unique ImmiRoot
which manages their interactions.

To support the xcode storyboard, a storyboard can be registered and used
to lookup and associate presenters to each other. This requires that the
ObjC class, to be constructed upon association, must be given a name
matching the Dart-side presenter name in the "storyboard id" field.

R=ager@google.com, zarah@google.com

Review URL: https://codereview.chromium.org/1150193003