Do not use intrinsics if their address is heap object tagged.
The address of the intrinsic is put in a table in the heap. That
doesn't work so well if it is heap object tagged.

This only seems to happen for me in one particular configuration:
debug_x86_asan. I haven't seen it in other configurations.

R=kasperl@google.com
BUG=https://github.com/dart-lang/fletch/issues/41

Review URL: https://codereview.chromium.org/1153733004