Standalone mock server.
This CL replaces the in-app mock server with a standalone off-device
mock server. The socket communication is abstracted to allow
"inverting" the client/server setup to support communication between a
developer machine running the mock server and an iOS device running the
app. In such a setup the device needs to establish the server socket.

R=ajohnsen@google.com
BUG=

Review URL: https://codereview.chromium.org/1165753007