Add temporary infrastructure to be able to reset the session state.
This is to allow the reuse of the VM for running multiple programs
in a row. Peter requested something like this as a stepping stone
towards being able to run tests by using the incremental compiler
to change the program instead of loading a new program.

R=ahe@google.com, ajohnsen@google.com

Review URL: https://codereview.chromium.org/1158903003