Presentation graph observers.
This CL adds a very simple form of mutation observation on the
presentation graph. Currently this allows reacting to parts of the graph
changing. It does not yet provide a description of what changed and is
not thread safe. Both issues will be addressed in subsequent CLs.

R=zarah@google.com
BUG=

Review URL: https://codereview.chromium.org/1173503005.