Fix casting of bit field on 32 bit systems.
The casting of pointers to integral types are not well-defined and are
implementation specific. It seems that most compilers agree that casting
a 32-bit pointer to a 64-bit integer should sign extend (whether or
not the resulting 64-bit integer type is signed). Avoid that by
casting to a word-size unsigned pointer type before extending the
amount of bits.

R=ajohnsen@google.com

Review URL: https://codereview.chromium.org/1193513004.