Add synchronization tokens for stdout and stderr when debugging.
When using the command-line debugger, output from the VM is delivered
through a stream. This means that output from the VM can be delayed
and it might not appear on the debugger stdout before the debugger
itself adds output on stdout (stack traces and other goodies). This
random interleaving of VM output and debugger information is confusing
and unexpected.

The only solution I can see is to put markers into the stdout/stderr
streams from the VM and make sure that we get those on all process
stop conditions before printing things from the debugger.

This patch set goes down that road but is incomplete. We cannot
guarantee that a newline is added to the VM stdout/stderr streams
and therefore the line-based checking of this CL will not work
in general. Before I got further down this route I wanted to get
your input and see if you have better ideas for synchronizing things?

This closes #54.

R=kustermann@google.com, ahe@google.com, ajohnsen@google.com

Review URL: https://codereview.chromium.org/1201393005.