New format for binary patches and objective-c patches.
This CL reimplements the binary format for patches. The format is then mirrored in objective-c such that updates on the UI can be done by interpreting the structure of patches. The interpretation of changes done to a node Foo is done by the methods -presentFoo and -patchFoo. Patches are applied using the methods -applyTo (side-effecting) and -applyWith (pure).

This CL also reorganizes the drawer/menu components to be compositionally defined using the new present/patch methods.

R=ager@google.com, zarah@google.com

Review URL: https://codereview.chromium.org/1211483002.