Debugger: hide all platform library frames by default.
This seems like the right thing to do. Users should be stepping through
their own code and not the library code. Using 't internal' users
can still step through the libraries if they want to.

This has the added benefit that we are not using main as the delimiter
of what to show. That doesn't work for async code.

R=ajohnsen@google.com, kasperl@google.com

Review URL: https://codereview.chromium.org/1203383002.