Final version, this cl splits the Foreign interface into a set of new classes: ForeignLibrary:   Holds a pointer to an open library. The existing solution did not allow opening   libraries not linked into the main program. We still allow using directly   linked in libraries by using ForeignLibrary.main.lookup ForeignPointer:   Holds a foreign pointer ForeignFunction:   Hold a pointer to a foreign function, this is a subclass of ForeignPointer,   and has all the functionality for calling functions.   We still need a more generic solution for calling functions with arbritary   arguments. ForeignMemory:   Holds a pointer to a piece of memory, and the lenght of the data. This is   a subclass of ForeignPointer. This can also be instantiated from dart using   the allocate and allocateFinalized.
This also adds some more testing, by adding a new c library that we compile and
use for validation of the functionality. This gives us the flexibility of
being more thorough in testing all our call and memory setting/getting functions.

Original working description for completeness.
Work in progres, please take a look and give early feedback if this is the way we want to structure this.

This is splitting the Foreign interface into several classes (maybe we should combine ForeignPointer and ForeignMemory).

It also adds some tests using a compiled c library that is part of this cl. We do this to actually test that we can use libraries not linked in and to better contoll what we get from there.

I still miss:
  Moving dart initiated allocation logic
  Remove old Foreign, and move the native functions up (but leaving them there for now allowed me to experiment with this before changing the IO logic)
  I have some more testing in mind.

BUG=
R=kasperl@google.com

Review URL: https://codereview.chromium.org/1209033003.