Implement a pub transformer for the immi compiler.
The transformer front-end implements the compilation of immi graph
nodes to Dart classes and generates the Dart implementation of the
fletch service. To do so, the code has be relocated to adhear to the
pub package structure and the various packages have been given a pubspec.

The idl and objc backends are still run with the old font-end and so
is the entire fletch service compiler. The remaining backends and the
sercice compiler will be replaced by a transformer in follow-up CLs.

R=ajohnsen@google.com

Review URL: https://codereview.chromium.org/1226483003.