Allow print interception and transmit VM output via the socket connection.
This is currently only used for the command line debugger. The other
parts of our infrastructure just throws the new commands away. That
will of course have to change in follow up changes.

I took the approach of using an explicit Print class with explicit
interception. The alternative would be to mock with stdout and stderr
directly. I found that to be too shaky and was usure if that would
work on all our platforms.

R=ahe@google.com, ajohnsen@google.com

Review URL: https://codereview.chromium.org/1248693002 .