Move c specific flags to common.gypi
Change out cflags section in common.gypi to cflags_cc, add a cflags_c section that only sets -fPIC.

We use this for our ffi_test_library (which we don't link in, it is simply a shared library that we use to validate ffi functionality).

Update the ffi_test_library target to not unset any flags, and to not do any defines.

@ahe, thanks for pointing out the cflags_cc and cflags_c solution.

R=ahe@google.com
BUG=

Review URL: https://codereview.chromium.org/1261003003 .