Simplify build setup for xcode development.
This CL reverts the use of transformers so that both immic and servicec
output all generated code to respectively an immi directory and a
service directory. For Dart, these two directories must be placed
on the package path. For xcode, these two directories are linked to via
Cocoapod.

R=zarah@google.com
BUG=

Review URL: https://codereview.chromium.org/1269563007 .