Get output from the VM printed in the fletch command console for the x-run verb.
Each verb has its own CommandSender and therefore we need to reattach
the ouput client for each verb that is interested in dealing with stdout
and stderr from the VM.

I have extended the session state with two buffering sinks that are
registered with the VM. These sinks buffer output from the VM until
a client is registered at which point all buffered output is delivered
to that client. While a client is attached the output from the VM
is directly delivered to that client.

R=ahe@google.com

Review URL: https://codereview.chromium.org/1259013004 .