Move toUpperCase and toLowerCase to pure Dart implementations
This is so that the tree shaker can omit them for programs that don't use them.
The toLower table is 500 bytes, the toUpper tables are 553, 172, 144 and 49
bytes where only the first is needed for regexps (which ignore multi-character
upper case rules). These tables are implemented with List<int> for now but should be something byte-sized.
BUG=
R=ahe@google.com

Review URL: https://codereview.chromium.org/1267163003 .