Simple regexp engine for fletch in pure Dart.
The intention is that this would be tree shakable, so you don't
pay for it if you don't use regexps.

The parser is recursive descent, and the byte codes are
built directly by the AST, without an intermediate represenation.
The "byte codes" are Dart integers in a list and the interpreter
is a switch-in-a-loop in Dart.

What works:
All parsing
+ ? * {} and non-greedy variants
Grouping with (?:), () and positive and negative lookahead
Character classes
Captures
Multi-line mode

What is missing:
Captures inside loops not quite right yet
Back-references
Case-independence /i flag
Detecting fixed width greedy quantizers to avoid linear memory
  use when matching large strings

BUG=
R=ager@google.com

Review URL: https://codereview.chromium.org/1258603002 .