Reland: Use full Session object for verbs and perform first simplifications.
A Session should not take the FletchSystem in its constructor. Instead
FletchDeltas are applied through a session which updates the system
to correspond to the last FletchDelta applied.

Also, an input line stream is only interesting if debugging. Therefore,
only pass it when debugging.

R=ahe@google.com

Review URL: https://codereview.chromium.org/1278493009 .