Test for servicec creating dir structure
For a servicec test to be successful the compiler needs to have produced
the correct directory structure for the specified targets. Currently,
the test is expected to fail, as the compiler does not have this
functionality yet.

R=zerny@google.com
CC=ager@google.com, zarah@google.com

Review URL: https://codereview.chromium.org/1287033011 .