Use sigkill on running process if it hangs after we did a sigterm on timeout.
Since the driver is catching the sigterm, and trying to close down the correct thing in the persistent process we may end up in a situation where we don't get shut down (we saw this on the bots). This would mean that we would hang until the bot times out and tear everything down.

BUG=
R=ahe@google.com, kustermann@google.com

Review URL: https://codereview.chromium.org/1302523002 .