Setup fletch in the github sample app for android.
In this setup the fletch source code is copied to Android Studio (by
using compile.sh) in order to make use of the new (experimental) native
code integration.  The native debugger works and can be used by
selecting the FletchDebug target/module before running. In addition
attachNativeDebugger should be set to true to allow time for the
debugger to attach.

R=zarah@google.com, ager@google.com
BUG=

Review URL: https://codereview.chromium.org/1283413004 .