Add servicec parser
This is a do-nothing parser, that just prints when it goes through its
different states. This is an intermediate stage before adding a listener
that can act on the switches between the stages and build an abstract
intermediate representation of the parsed IDL file.

Review URL: https://codereview.chromium.org/1305823002 .