Split listener; simplify errors; add 2 tests
The servicec Listener is now a base class that just forwards the input
tokens to its methods and keeps a track of the errors that occurred
during parsing. There is also a DebugListener class which can attach to
objects that implement the Listener interface and print debugging
information about their operation.

The servicec compiler errors were simplified from a a class hierarchy to
an enumeration, but will most likely get more complicated, once we
figure out what is the best way to implement them. For now, we Keep It
Simple.

Two tests are added that test the error reporting and error recovery
capabilities of the servicec Parser. One addresses recovery from
syntactically incorrect function list inside a service declaration and
the other tests for recovery of missing closing } at the end of a
service declaration. They are not passing yet, since we don't have the
required functionality.

R=zerny@google.com

Review URL: https://codereview.chromium.org/1305123004 .