Add rudimentary unmatched { recovery listener
The new class detects top-level keywords which are used inside other
top-level declarations and assumes that the reason for this is an
unmatched {. It injects an UnmatchedToken in the token stream which is
propagated until it is skipped by the parser where normally a } is
expected. This allows the parser to recover from such errors.

BUG=
R=zerny@google.com

Review URL: https://codereview.chromium.org/1316993002 .