Reland: Use minimal double-conversion library.
This cuts 16kB off the fletch-vm executable. Double conversions will be slower but still correct.

This change creates a temporary copy of double_parse_test.dart without
the cases that fail because of strtod on MacOS. That way we do not
lose test coverage. Once we land changes in the Dart SDK repo
to enable us to mark only parts of the test as failing we should
remove the temporary copy. That is tracked in

https://github.com/dart-lang/fletch/issues/114

R=floitsch@google.com

Review URL: https://codereview.chromium.org/1319673004 .