Github sample: use the non-inverted github mock server.
The inverted server does not work properly because of possible races
between calls to Socket.accept.

For iOS, the mock server does not work.
For android, the mock-server port should be forwarded with:
  $ adb reverse tcp:8321 tcp:8321
and the mock server can be started just once before starting the app.

R=zarah@google.com
BUG=

Review URL: https://codereview.chromium.org/1321103003 .