Implicit session
If one refers to a session that doesn't exist, a session with that name is created.

When a session is created, it becomes the default session for subsequent commands.

If no session is specified, and there's no default session, a session named "default" is created.

R=ager@google.com

Review URL: https://codereview.chromium.org/1324423002 .