Two small bugfixes and adding support for passing an immutable context.
This fixes a bug/type in the typed data fromList constructor and a bug in the file open method related to creating a new file with the right permissions.
Finally it adds support for passing an immutable context  to the process spawned from the Socket.spawnAccept method.

R=kustermann@google.com, sgjesse@google.com, ager@google.com, ajohnsen@google.com
BUG=

Review URL: https://codereview.chromium.org/1335573003 .