Don't make cross-platform field offset assumptions
Until now we assumed that the field offsets within C++
objects were the same on the host (generating the
interpreter) and the target (running the VM).  This is
not always the case, which caused the random number
generator change to fail. This CL decouples those
offsets partially and asserts that we use the same
offsets on host and target.

We could do the same for LookupCache::Entry, but
for now that's a pretty simple layout and I don't
expect deviations.

static_assert is a C++11 feature.

R=ager@google.com
BUG=

Review URL: https://codereview.chromium.org/1331803004 .