Add package with I2C support for Raspberry Pi
This includes a device interface for HTS221 humidity and temperature
sensor and the LPS25H preassure sensor on a Raspberry Pi Sense HAT.

The implementation experiments with having the actual I2C bus access
code running in a separate process. When the I2C API is used a message
(containing a closure to run) will be sent to the process hosting the
I2C bus and a the result returned through a port.

This is not realy isolated yet as the process hosting the I2C bus
still has full system access and is currently handed a file descriptor.

Testing is currently very limited.

R=karlklose@google.com, kustermann@google.com, sigurdm@google.com, kasperl@google.com
BUG=

Review URL: https://codereview.chromium.org/1340533002 .