First version of the fletch agent.
The agent is meant to make it more seamless/simple to deploy new fletch-vm instances. For the Raspberry Pi2 it will be installed on the sd card via a script after which the agent will run on bootup.
This allows the fletch compiler to spawn new fletch-vm instances. The idea is the agent will report back which pid is spawned and more importantly what port it is listening on.

R=sgjesse@google.com, ager@google.com, lukechurch@google.com, pahe@google.com, zerny@google.com
BUG=

Review URL: https://codereview.chromium.org/1331833004 .