Add scripts for configuring a raspberry pi image with fletch.
This includes a few other improvements as well.
The agent is driven by environment variables (in lack of command line args). It now printing its log and pid file in the correct directories.
Also fixes a bug in the file open method to use octal 0666 instead of decimal:P
Finally added a ForeignCString to the ffi interface. This is primarily used to convert a cstring into a Dart string.

R=sgjesse@google.com, ahe@google.com, karlklose@google.com
BUG=

Review URL: https://codereview.chromium.org/1355823002 .