Github sample: Mirror sliding-window patch structure from Android in ObjC.
This change fixes present to no longer assume an empty sliding window
and makes it possible to both add and remove items from the item list
represented by the sliding window. It is still not reliable to switch
between two distinct lists because the scroll position is not properly
tracked by the presentation graph, but with this CL a change will adjust
to an approximate location based on where the sliding window location.

R=ager@google.com, zarah@google.com
BUG=

Review URL: https://codereview.chromium.org/1359633002 .