Allow inspection of constants without marking the constant values as used.
This allows the code generator to inspect constants to make code
generation decisions without signing up to using the constant values
in the generated code.

I encountered this while working with bigints. If I compile without
bigints the system would still attempt to transmit a bigint constant
to the VM eventhough I do not use it in the generated code.

R=johnniwinther@google.com, ahe@google.com

Review URL: https://codereview.chromium.org/1359333002 .