Add bigint implementation.
The implementation is based on Dart VM pure Dart bigint implementation.
The first patch set contains a pure copy which means that you can
see the changes I did in patch set 2 for that file.

Currently, bigint support is disabled with a compile-time constant.
Once the setting file setup lands I will flip the flag to true
by default and we can disable bigint support for smaller devices
where we do not want to pay for them in the snapshot. Currently
the cost for bigints is around 17kB in the snapshot.

R=erikcorry@google.com, ajohnsen@google.com, kustermann@google.com

Review URL: https://codereview.chromium.org/1364723002 .