Enable bigint implementation by default.
It is controlled by a constant in the settings file. Here is how you
create a snapshot with and a snapshot without support for bigints
given the checked in settings file and a copy where the only
change is that the value of the constant is changed to "false".

  $ ./out/ReleaseIA32/fletch create session bigint
  $ ./out/ReleaseIA32/fletch export hello.dart to hello.snapshot.bigint
  $ ./out/ReleaseIA32/fletch create session no-bigint with .fletch-settings-no-bigint
  $ ./out/ReleaseIA32/fletch export hello.dart to hello.snapshot
  $ ll hello.snapshot*
  -rw-r----- 1 ager eng 17995 Sep 23 17:18 hello.snapshot
  -rw-r----- 1 ager eng 34801 Sep 23 17:18 hello.snapshot.bigint

R=ahe@google.com, erikcorry@google.com

Review URL: https://codereview.chromium.org/1361973003 .