Add newline tokens to servicec scanner
The new lines are used in error recovery: when there is a bad
declaration the line where that declaration starts is ignored. The
testfile newline.idl demonstrates the improvement: without it, there is
a single error that is reported; with the change, there are two errors -
one per bad declaration.

BUG=
R=zerny@google.com

Review URL: https://codereview.chromium.org/1364683005 .