Add script for creating an sdk bundle This is dependent on: https://codereview.chromium.org/1374163003/
I originally a large set of the lists of things to move/copy at the top of the file, but with some of the special casing it actually made it more readable having it in place in the functions. I am happy to move more stuff out of the functions if requested.

I will upload testing support in a seperate cl to not have to big a change.

R=sgjesse@google.com, wibling@google.com
BUG=

Review URL: https://codereview.chromium.org/1372083004 .