Reduce possible spinlock contention during Port.send()
Spinlocks should be held for extremely short amounts of time. Memory
allocation can cause syscalls and which can take for potentially 10s
to 100s of us.

It's also nice to reduce the amount of time we hold the lock so it becomes less
contented if multiple processes send to the same port at the same time.

R=kasperl@google.com

Review URL: https://codereview.chromium.org/1375213002 .