Github sample: Disable rotation when a list item is selected.
This change is a temporary fix to avoid the issue of reestablishing the
UI state to a selected position which is complicated by two issues:
1. the selected state of the sliding window is not present in the
sliding window itself; 2. on restore Android will implicitly recreate
views that we should rather manually create on present/patch.

R=zarah@google.com
BUG=

Review URL: https://codereview.chromium.org/1375293002 .