Fix print interceptor bug.
This fixes a small bug where we always called the first print interceptor in the list instead of the current. It caused us to print output as many times as the number of interceptors (which is 1 by default and 2 if using the --log-dir flag).

R=ahe@google.com, kustermann@google.com, ajohnsen@google.com, sgjesse@google.com
BUG=

Review URL: https://codereview.chromium.org/1374663004 .