UncaughtExceptions Part 1: Do not unwind coroutine stacks in case of uncaught exceptions, let debugger handle uncaught exceptions
Previously we would unwind all coroutine stacks except the last one, even if the
exception is not actually caught. This leaves the stack of a process in a state
where one cannot even inspect the state of of exception throw time any more.

This CL also makes the VM only print an uncaught exception message to stdout if
there is *no* session attached, or the session is not able to handle it. The
rational behind this is that the session can provide more context than the VM
in case of an uncaught exception and can produce much nicer error messages.

=> If a debugger/session is attached, it must take care of inspecting the
   stack/exception and presenting it in a nice way to users.

This is Part 1 for getting nice uncaught exceptions messages.

BUG=https://github.com/dart-lang/fletch/issues/80
R=ager@google.com

Review URL: https://codereview.chromium.org/1375633005 .