Add cyclic struct import detection in servicec
A single error is reported per node if there is one or more previously
unseen cycle reachable from that node. This will lead to different
amount of errors depending on the order of traversal of the nodes, as in
the following example:

       .---.             .---.
      v     |           v     |
A --> B --> C --> D --> E --> F

If we traverse the nodes in alphabetical order we will get 1 error,
since the whole graph will be traversed, just by starting at A (and
there is a single error per node that can reach cycles.) However, there
will be 2 errors if we traverse in reverse alphabetical order, since
only E and F will be initially traversed. Then D will be traversed and
no error will be reported (the cycle of E is already seen). C and B will
be traversed and an error will be reported. And finally, A will be
traversed and no error will be reported.

We can change that later, if we think it is important to have a
different amount of errors than currently reported.

BUG=
R=zerny@google.com

Review URL: https://codereview.chromium.org/1383993002 .