Move struct cycle detection to a separate file
As opposed to the message in the previous comments, now there is an
error reported for a node participating in a cycle. It still depends on
the execution order how many errors will be reported, as in the
following example:

.-----------------.
.-----------v     v
A --> B --> C --> D
^-----------'     |
^-----------------'

If the path A - B - C - D - A is traversed first, then a single error will
be reported. If A - D - A and A - C - A are traversed first, then two
errors will be reported.

However, we are not concerned with this for now, since the current
solution is good enough. Ideally, we will report the link that is
causing the problem, but this is hard to find and it doesn't outweigh
the effort.

BUG=
R=zerny@google.com

Review URL: https://codereview.chromium.org/1376513004 .