Fix race condition in shutdown procedure
The main scheduler thread will shutdown - possibly stopping the program and
deleting all processes - once it gets a signal to do so.

This "signal" can be for example:
  * the number of processes reached 0
  * a process had a fatal error (e.g. uncaught exception) and shutdown_ was set

It is safe if a scheduler worker thread sends this "signal", since the main
thread will get the worker threads to stop first.

But if a session thread does this from the side, it is important to first delete
a process and then send the signal atomically. Once the signal is sent, the
scheduler might be already dead and can therefore no longer be used.

A cleaner solution would be to simply not allow deleting processes from the
side, but rather let the normal scheduler worker threads handle it.

BUG=https://github.com/dart-lang/fletch/issues/185

Review URL: https://codereview.chromium.org/1395063002 .