Run the service compiler at test time
This CL adds a rule for running the service compiler at test time. The
service compiler needs to be run and its output is used to build
respectively C/C++ and Java executables. These executables can then be
run to preform the actual tests.

Currently we only use the C/C++ generated output to build and execute a
test. The Java tests are not yet automated.

BUG=
R=ahe@google.com

Review URL: https://codereview.chromium.org/1408803005 .