Treat dart keywords as identifiers in servicec
The dart keywords which are not servicec keywords (all of them for now,
but not necessarily so in the future) are converted to StringTokens with
kind IDENTIFIER_INFO. This is later used to validate that these tokens
are identifiers (in contrast, '[', '{', etc. are not, as their kind is
not IDENTIFIER_INFO).

BUG=
R=ahe@google.com

Review URL: https://codereview.chromium.org/1408303003 .