Fix issue with flags and errnos on Mac being different than linux.
The reason the agent_tests failed was due to the agent failing to create a log file since the O_CREAT flag has a different value on Mac. While at it I found a number of other flags that are different. One O_NONBLOCK caused the samples github_api_tests to fail once it was initialized with the correct O_NONBLOCK flag due due to an errno being different as well.

R=ager@google.com, sgjesse@google.com, ajohnsen@google.com
BUG=

Review URL: https://codereview.chromium.org/1425033005 .